# perf(assump): add negated items to the values of get_known_facts_dict()