--- a/sympy/assumptions/facts.py
+++ b/sympy/assumptions/facts.py
@@ -185,7 +185,7 @@ def compute_known_facts(known_facts, known_facts_keys):
 
 
 def single_fact_lookup(known_facts_keys, known_facts_cnf):
-    # Compute the quick lookup for single facts
+    # Return the dictionary for quick lookup of single fact
     mapping = {}
     for key in known_facts_keys:
         mapping[key] = {key}