--- a/sympy/assumptions/facts.py
+++ b/sympy/assumptions/facts.py
@@ -193,6 +193,8 @@ def single_fact_lookup(known_facts_keys, known_facts_cnf):
             if other_key != key:
                 if ask_full_inference(other_key, key, known_facts_cnf):
                     mapping[key].add(other_key)
+                if ask_full_inference(~other_key, key, known_facts_cnf):
+                    mapping[key].add(~other_key)
     return mapping