<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->


#### Brief description of what is fixed or changed

`_ask_single_fact()` function is introduced for cleaner code.
More items are added to `get_known_facts_dict()`. This improves the speed of some queries.

In master:
```python
In [1]: %timeit ask(Q.even(x), Q.odd(x))
1.75 ms ± 18.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

With this change:
```python
In [1]: %timeit ask(Q.even(x), Q.odd(x))
620 µs ± 1.82 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

#### Other comments

Please note that this adds items only to the single fact lookup dictionary. It does not complicate `satask`.

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
NO ENTRY
<!-- END RELEASE NOTES -->